Server/requirements.txt
faker
django  
djangorestframework
django-cors-headers
django-environ==0.12.0
psycopg[c]==3.3.2
argparse
pydantic